Some rust project conveniences & repo consistency (#1)
* Some rust project conveniences & repo consistency

Most notably, added a LICENSE file with the BSD-2-Clause-Patent
licence in it, with the four of us as copyright holders. Please
make sure I spelled your name correctly, and also please give me
a thumbs up or down on that license choice. I believe it ticks the
boxes we discussed this morning.

Stuck in a placeholder for design docs & usage docs, as well as
an opinion in the readme about where *implementer* docs should live.

* Update our Cargo.toml and the license checker

... with our actual license.