Reorg to a workspace with three sub-crates
The `engine` crate is a library with the embeddable engine
with our glue around it. This still uses anyhow for errors, though
we should eventually replace it with custom errors using `thiserror`
or something else suitable for libraries.

The `once` subcrate is the existing cli that runs the wasm payload
you point it to then exits, but reworked to use the engine glue
instead of using wasmtime directly.

The `daemon` subcrate is a scaffold for an http server that accepts jobs
via a POST endpoint. Well, it will eventually do so.