Refactor JobQueue to avoid global state
Rather than using a global singleton Vec<Job> as the job queue with a bunch of functions that
manipulate the global singleton, this makes a JobQueue struct and moves all of the functions
onto it.

Having done that, the JobQueue ended up fitting into Axum's state nicely. This is much better
in prett much every dimension and is fairly obvious in retrospect; I am not sure why I did it
wrong the first time around. 😅