Checkpointing a POUNCE
The daemon now accepts binary payloads, runs them, and responds with
the results. There is a new CLI that exercises this behavior.

The engine runs the wasm immediately and does not queue it for
processing, but at least it's bootstrapped.

Next up: porting over the code from meshy-mesh to find our job
queue servers!