Rust Zamboni mode: lints, license checks
Add a license clarification for `ring` to the cargo-deny setup,
which allows us to turn on that check in CI again. Also, add the
picky lint setup to the heads of more subcrates, and fix the
resulting warnings.

I love the word "lint" for this. Programming jargon has such
fun wordplay and metaphor in it.