Start a long journey with CLI output design
Pounce now has some of the tools we'll need to do good CLI output
design over time.

- Added `owo-colors` to colorize terminal output, and used it in
  a couple of println!s as an example.
- Added `loggerv` to provide colorized human-intended log macro
  output, controllable via a verbosity level.
- Added a counted `-v` flag to pounce and fed it to loggerv.
- Took all development or debugging prints and moved them to
  log macros, so they can be silenced if somebody wants to run the
  cli without getting details.

In addition, moved mdns-sd to a workspace-level dependency and turned
off its logging feature for all consumers. I reason that we want to
make sure we're using the same version of that library in all of our
components.