Add a starting point for serval error types
Introduced the crate `thiserror` for deriving custom error
types from strings and other custom error types, and put it
to work in our utils library. Note that no changes are needed
in code that consumes the library, because of the glue from
`anyhow`. This pattern lets us start to be able to handle errors
less generally in application layers, if we should need to.

Did a little gardening in the library to pull some related things
into single files. We can always split them out again later if they
grow.