The content-addressed storage is now a tiny lib.
It's now in utils::blobs. The API is `store()`, `get_stream()`, and
`get_bytes()`. Made castaway use the library, which reduced it to pure
http setup and response-building.

I took castaway's endpoints and copied them wholesale into
serval-agent. The agent's versions of the endpoints are `GET
/blobs/:id` and `PUT /blobs` because everybody writes http endpoint
names differently and I pluralize the nouns. This is not meaningful,
to be clear.

MORE EXCITINGLY, serval-agent also got a `GET /run/:blob_id` endpoint
which probably really shouldn't be a GET but I'm making this PR fast
instead of fixing that. This runs a stored binary! No input yet,
because again, PRing for discussion first, and I think that requires a
little thoughtful design first.