serval-agent - allow an instance to run without offering blob storage
Right now, our blob storage is a temporary implementation that just stores data in $TMPDIR. If
multiple copies of the agent are running on different machines, that means that blob storage
won't be reliable -- if you don't happen to ask an instance on the correct box, you might not
get your data back.

To avoid this, let's make it so you can run an instance of the agent without it implementing
any blob storage stuff. Changes required to do this:

1. Add command line parsing via clap and allow people to specify whether this instance should be
a storage node or not. Rather than a true/false switch, I've made this auto/never/always, with
the assumption that auto will someday make some judgements about whether the current machine is
a viable platform upon which to run the storage system. e.g. an embedded device without much
disk space might decide not to run the storage server.

2. Refactor `RunnerState` so that it holds an `Option<BlobStore>` instead of a `BlobStore`. This
involved some yak-shaving to make the ::new function return a `Result`.

3. Change blob-related API endpoints to return 503 errors if you try to hit them when storage is
disabled.

Future changes required but not done here:
- API calls to /blobs/... should proxy to a node that is advertising the serval_storage role,
rather than just returning a 503 error.

While I was was in here, I modified `BlobStore::new` function to ensure that the path it has been
given exists and points to a writable directory. If the path does not exist, we try to create it.
I also changed the default path from `/tmp` to `/tmp/serval_storage`.

Phew!