Support for WASM extensions (formerly known as capabilities)
If you start the agent up with EXTENSIONS_PATH set, it will enumerate all of the .wasm
files in the given location and make them available as imports for all jobs that are
run on this node. The base filename (without the extension) is used as the module name,
so if you have a file `foo.wasm` that exposes a function called `bar`, you could make
use of that from within a Rust WASM file with a declaration like this:

````rust
/**/#[link(wasm_import_module = "foo")]
extern "C" {
    fn bar(a: u32) -> u32;
}
````

There are plenty of considerations left unconsidered for Future Us, in particular
around permissions and whether every job should actually be able to make use of every
extension. (e.g. Perhaps some extensions should only be available to jobs associated
with a particular namespace, once we implement namespaces.)