Clean up the task manifest a bit.
This needs more than the seat-of-the-pants design I've given
it just now, but this pass removes redundant information.
Also, we now standardize on using the fully-qualified name
everywhere.

While I was there, introduced some `prettytable` to the CLI
to start formatting some displays nicely. This is throwaway
work in the long term, but it makes things a little nicer now.