Refactor the proxy mount routing extensively (#42)
Each`v1` api submodule now exports `mount()` and `mount_proxy()` as the functions to call to mount all their routes. This corrals all knowledge of exact endpoint paths to the modules, so the details of the path are no longer something `main()` knows about. Rewrote the middleware to be handlers mounted with the `any` verb. There's now a little code duplication in each of the proxying endpoints, but this does not yet hit my threshold for generalizing it. I do like that this approach let me hide all of the endpoint handlers inside their modules.

Additionally, removed all dependencies on native-tls from our usage of reqwest. In theory, this will make us more easily portable. In theory.