Start emitting metrics. (#44)
This commit adds the `metrics` crate, which is a commonly-used set of
macros for emitting metrics used by many output generators, similar
to the `log` and `tracing` crates. Typical metrics look like this:

```rust
metrics:increment_counter!("process:start");
metrics::histogram!("memory", someNumberHere);
metrics::increment_gauge!("gauge_name", someNumberHere);
```

This is literally all you need to do to send a new metric. The
crate offers more affordances if you need them or want to re-use
a specific metric emitter in a function.

Added basic metrics to each of the endpoints in the agent, as well as
some process restart metrics.

Added one output sink, which we'll want to replace immediately because
it's only suitable as a hello-world proof that metrics are being sent.
In order to see metrics from the process, you'll need to check out
my fork of the metrics-rs repo and run the `metrics-observer` subproject.

```terminal
gh repo clone ceejbot/metrics-rs-fork
cd metrics-rs-fork/metrics-observer
cargo run -- 0.0.0.0:9000
```

Then run the agent and do some things. This is obviously unacceptable
long-term, but we'll pick a first real candidate for metrics output
eventually. In the short term, it's good to get into the habit of
instrumenting code as we write it.