Switch from mdns to using the kaboodle peer mesh (#48)
The agent and the cli now both attempt to join a network of peers, established via kaboodle, and use that network instead of mdns to discover relevant services. 

The most interesting implementation details are in the file `utils/src/mesh.rs`. This file includes some wrapper traits that probably belong in Kaboodle itself, since they were designed to be re-usable and to encode the responsibilities of any Kaboodle implementation: KaboodleMesh and KaboodlePeer. There are implementations plus some serval-specific functions in the structs ServalMesh and PeerMetadata. (All names are negotiable. I already don't like my initial choices.)

The peer metadata struct gets a little complicated because I attempted to provide some future-proofing via a way to version it. It's encoded via `binencode`, with a version envelope wrapper to allow us to make newer agents that change what gets packed into identity able to interoperate with older agents, and let older agents ignore newer ones. I welcome improvements to this somewhat ad-hoc design.

Agent roles are now an enum, `ServalRoles`. The `strum` crate provides Display and FromString implementations. This adds some safety and validity-checking to what had been some magic strings.

We knew we'd stop blocking in the cli some day, and this is that day: the cli is now fully async.

Finally, and most noisily, corrected the spelling of "Wasm" everywhere. Apologies for that.