Checkpoint: added a `monitor` command to the cli
This command tells the CLI to join a mesh and print out
joins and leaves as it is informed of them.

The CLI and the agent both now respect a new env var: MESH_INTERFACE.
If you don't provide one, kaboodle defaults to the best it can find.
If you do provide a string here, we find an interface with a matching
name. This allows me to specify ipv4 and therefore run tests successfully
locally.

The CLI now only connects to a mesh to get a base url if it ever needs
a base url. It doesn't need a base url for this command, for instance.
This change caused a lot of trauma in the cli as things became async
that were not previously so. tl;dr the CLI now uses async_once_cell.

Generalized some other things as needed. Stole some network interface
selection code from kaboodle itself. (We might want to have kaboodle
take an interface suggestion via string; worth thinking about anyway.)