Mesh shenanigans in the CLI (#53)
Three changes, two notable!

1: Added a new `monitor` command to the cli. This command joins the
cli to the mesh as a `client` role, and prints out joins and leaves
as it is informed of them.

The CLI and the agent both now respect a new env var:
`MESH_INTERFACE`. If you don't provide one, kaboodle defaults to the
best it can find. If you do provide a string here, we find an
interface with a matching name. This allows me to specify ipv4 and
therefore run tests successfully locally.

The CLI now only connects to a mesh to get a base url if it ever needs
a base url. It doesn't need a base url for this command, for instance.
This change caused a lot of trauma in the cli as things became async
that were not previously so. tl;dr the CLI now uses `async_once_cell`.

Generalized some other things as needed. Stole some network interface
selection code from kaboodle itself. (We might want to have kaboodle
take an interface suggestion via string; worth thinking about anyway.)

2: The CLI now uses the single-peer discovery affordance in the latest
Kaboodle. This speeds up all CLI operations _considerably_.

Added new endpoints to the agent to support requesting and receiving
lists of peers via the HTTP api. Wrapped kaboodle's new
discover-a-peer api in a new `mesh::discover()` function that reuses
our env var hinting code. Reworked the CLI's http base url discovery
code to use the the new `discover()` function. This code simplified a
lot!

`PeerMetadata` and `ServalRole` now both are serializable via Serde.

3: Nuked strum. Strum is nice and it does handy things, but we were
only using it for a fairly trivial-to-implement display trait for
ServalRole.