KaboodlePeer - replace Option<SocketAddr> with IpAddr
Having `PeerMetadata` (which implements `KaboodlePeer`) take an `Option<SocketAddr>` was awkward for
two reasons:

1. We don't know our SocketAddr until we actually join the mesh -- and it will change if we `.stop()`
and `.start()` again later.

2. Every peer other than ourselves will definitely have a `SocketAddr`, so having it be wrapped in
an `Option` is a bit awkward -- we have to write code to handle the `None` case  even though that
will never happen for anyone other than us.

Replacing this field with an `IpAddr` solves both of those problems. We lose the ability to send a
message to our peers over their Kaboodle port number, but that's a feature rather than a bug in my
opinion.