Checkpoint: a serval agent api client
There's a client sub-crate that provides a client for the serval agent
API. It hides the choice of reqwest as an http client-- mostly-- and
lets you call async functions to get data.

Re-implemented the cli commands using the api client. While I was
there, removed cli support for dead api endpoints and added support
for the ones we actually have implemented.

The api client is not yet pulling its weight, IMO, but it will as soon
as I make the storage proxy start using it. Oh yeah, this checkpoint
also includes work towards letting job runners be able to operate
without local storage-- the client will let them call out to a peer
that *does* have storage.

Moved storage stuff out of util and into the agent. I'm not yet sure
about this location, so it might move again.