Make jobs runnable even without local storage (#57)
Job runners no longer need local storage. A node with the runner role
but without storage now reaches out to a peer with storage to get the
executable it's been asked to run.

CAVEATS: There is no attempt made YET to see if a peer node actually
has the items we're asking for. So you might try to run a job and hit
a peer that doesn't have the thing. We have yet to design how
distributing storage will work, so I haven't touched this problem yet.

Here's how this PR achieves this:

The repo has grown a new api-client subcrate. This implements an http
client for the serval API. It hides the choice of `reqwest` as an http
client-- mostly-- and lets you call async functions to get data.

Re-implemented the cli commands using the api client. While I was
there, removed cli support for dead api endpoints and added support
for the ones we actually have implemented.

Moved the storage implementation out of utils and into the agent
sub-crate, in its own sub-module. The storage implementation is now a
trait plus two useful structs that implement the crate. The trait is
`RunnerStorage`. This captures what a job runner needs from storage,
plus some other not-yet-needed but easy-to-achieve fetches. 

The `cacache`-backed storage struct, `BlobStore`, now implements this
trait plus some functions that cacache makes easily available.
Notably, getting executable bytes as a *stream* is in the api for this
struct and *not* for the trait. I was stuck on wrangling AsyncRead
traits, then realized that the job runner doesn't need this, so I
punted on it for now.

The second implementor of the RunnerStorage trait is `StorageProxy`.
This uses the API client to make requests from a mesh peer that
advertises the storage role.

The job runner now uses a lookup function to get its storage
implementation. If the node has storage, it snags the `cacache`-backed
one it created at startup. If not, it looks up a fresh peer that
advertises the role, makes an api client via a StorageProxy, and asks
that peer for the data it needs.