S3 buckets now support fetching content by address (#63)
Added a layer of indirection to S3 storage as a cheap index. Instead of writing data to the user-provided key, we write an object containing the integrity hash for the data. We then create an object that contains the data using the integrity hash as a key. Re-implemented all data stores and fetches in the bucket via this indirection. This is of course invisible to clients of the storage layer.

Implemented `HEAD /v1/storage/data/*address` via `has_content_address()`, which is now possible that the bucket storage supports it. Also implemented `POST /v1/storage/data` for storing data without any human-centric key.

The term sri is now gone from all API surfaces. Instead we use "integrity" as the word for this concept.

Improved some error reporting, and fixed a bug with how the cli reported hashes. (It was reporting hashes of the hashes instead of printing the hashes, because I held the API wrong.)